REF: formatting refactors & improvements to scatterplot spec
