+++ q2_vizard/__init__.py
@@ -7,11 +7,10 @@
 # ----------------------------------------------------------------------------
 
 from ._version import get_versions
-from ._heatmap import plot_heatmap
-from ._scatterplot import scatterplot_2d
-from ._util import json_replace
+from .heatmap import plot_heatmap
+from .scatterplot import scatterplot_2d
 
 __version__ = get_versions()['version']
 del get_versions
 
-__all__ = ['plot_heatmap', 'scatterplot_2d', 'json_replace']
+__all__ = ['plot_heatmap', 'scatterplot_2d']
