+++ q2_vizard/heatmap.py
@@ -12,7 +12,7 @@ import pkg_resources
 import jinja2
 import json
 
-from q2_vizard._util import json_replace
+from ._util import _json_replace
 
 
 def plot_heatmap(output_dir: str, data: pd.DataFrame, transpose: bool = False,
@@ -49,11 +49,11 @@ def plot_heatmap(output_dir: str, data: pd.DataFrame, transpose: bool = False,
     else:
         order = False
 
-    full_spec = json_replace(json_obj, data=data, x_label=x_label,
-                             x_label_name=x_label_name,
-                             y_label=y_label, y_label_name=y_label_name,
-                             title=title, measure=gradient,
-                             measure_name=measure_name, order=order)
+    full_spec = _json_replace(json_obj, data=data, x_label=x_label,
+                              x_label_name=x_label_name,
+                              y_label=y_label, y_label_name=y_label_name,
+                              title=title, measure=gradient,
+                              measure_name=measure_name, order=order)
 
     with open(os.path.join(output_dir, "index.html"), "w") as fh:
         spec_string = json.dumps(full_spec)
